requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
numpy==2.1.1

httpx==0.27.2
pymongo==4.8.0
groq>=0.12.0
python-dotenv>=1.0.0
